requirements.txt
# Original dependencies
streamlit==1.28.1
flask==2.3.3
pymongo==4.5.0
pandas==2.1.1
requests==2.31.0
python-dotenv==1.0.0
plotly==5.17.0
streamlit-option-menu==0.3.6
openai==1.3.0
PyPDF2==3.0.1
openpyxl==3.1.2
Pillow==10.0.1
pytesseract==0.3.10
pydantic==2.4.2

# ML and AI dependencies
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.24.0
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.10.0

# Qwen2-VL specific dependencies
qwen-vl-utils>=0.0.3
tiktoken>=0.5.0

# Additional ML utilities
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For GPU acceleration (uncomment if you have CUDA)
# torch-audio>=2.0.0+cu118
# torchvision>=0.15.0+cu118